Channel/requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.1+computecanada
idna==3.4
kiwisolver==1.4.5+computecanada
matplotlib==3.6.1+computecanada
numpy==1.26.4+computecanada
packaging==21.3
Pillow==10.0.0+computecanada
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1+computecanada
requests==2.28.1
scipy==1.10.1+computecanada
six==1.16.0
tqdm==4.64.1
typing-extensions==4.3.0
urllib3==1.26.12
torch==2.1.1+computecanada
torchmetrics==1.4.0.post0
torchsummary==1.5.1
torchvision==0.16.1+computecanada
scikit-image==0.24.0+computecanada
lpips==0.1.4+computecanada